+++ acm_web_site/apps/business/web_site/api/team_service.py
@@ -1,4 +1,4 @@
-from web_site.models import Team
+from acm_web_site.apps.business.web_site.models import Team
 
 
 def get_teams():
