+++ acm_web_site/apps/business/web_site/serializer/tutorial_serializer.py
@@ -5,7 +5,7 @@ import base64
 from rest_framework import serializers
 
 from acm_web_site.settings import MEDIA_ROOT
-from web_site.models import Tutorial
+from acm_web_site.apps.business.web_site.models import Tutorial
 
 
 class TutorialSerializer(serializers.Serializer):
