+++ acm_web_site/settings/__init__.py
@@ -1,0 +1,1 @@
+from .common import *
