+++ acm_web_site/urls.py
@@ -5,6 +5,6 @@ from markdownx import urls as markdownx
 
 urlpatterns = [
 	path('admin/', admin.site.urls),
-	path('', include('web_site.urls')),
+	path('', include('apps.business.web_site.urls')),
 	path('markdownx/', include('markdownx.urls')),
 ]
