+++ acm_web_site/wsgi.py
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "acm_web_site.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "acm_web_site.settings.development")
 
 application = get_wsgi_application()
