+++ package.json
@@ -4,7 +4,7 @@
   "description": "Website API for Javeriana's ACM chapter",
   "scripts": {
     "requirements": "pip install -U -r ./requirements.txt",
-    "start": "python manage.py runserver"
+    "start": "python acm_web_site/manage.py runserver"
   },
   "keywords": [
     "pyhton",
